Sort out board settings, mostly regarding variation display
Instead of the "small stones" option which didn't work, we now allow
letters as variation markers.  Together with the option to mark children or
siblings, that means we can really get rid of the mark brothers/mark sons
menu items.

Make the preferences dialog a little less confusing by using combo boxes in
more places instead of a series of checkboxes.

Allow the user to specify how to handle the SGF ST (style) property which
controls variation display.