Fix handling of board wood images
First, make a combo box to make sure all the five available images can be
selected, as well as a custom filename.  Make the filename selection stay in
the same directory as the previous file if possible.

Then, move the loading code to settings, since imagehandler is per-window
and should not try to manage a single pixmap that is identical for all.

While there, also remove the unused altGhost pixmaps to further simplify the
code.