Rewrite QGtp again to make it work on Windows
QProcess and its communication channels behave differently on Linux and
Windows.  On the latter in particular, it seems to be impossible to use
both waitForReadyRead and the readyRead signal alternately, it seems it
has to be one or the other.  But there were occasional strange issues
on Linux as well.

This rewrite changes everything in QGtp to use the signals, buffering
up output and distributing it to our own receiver queue once a full
line has been read.