Reimplement clock displays, and warn when clock times run low
This restores some missing functionality by playing sounds when times get
low in a match game.  Also, the clock display now flashes red as well.

The clock display is now a ClockView, derived from QGraphicsView.  The idea
was to have it just as a black bar with white text (and color changes when
warning), and there are some discussions found on-line about how getting
push buttons to look exactly right is not completely portable.  This way,
we have full control.