Tidy up font saving/restoring
QFont has toString and fromString functions explicitly documented to be
useable in saving/restoring settings, so use these, replacing code with
platform #ifdefs.  Keep fontToString around, but make it show a limited
amount of information for use in the font selection buttons.

Also ensure we set the console font correctly in mainwin.