Add more internal self-tests, and a way for SGF files to report nonfatal errors
Kogo's Joseki Dictionary for example plays a suicide move, followed by a
stone on another stone.  These changes make us deal with this gracefully,
truncating the loaded file at the invalid position and reporting an error
through a message box but still allowing the file to be loaded.

The self-tests found an issue where we'd recalculate liberties, overwriting
-1 markers we set earlier.