Tidy up on-line game end code
Introduce new game_end methods that are used instead of signal_gamemove,
which is now used only for games in progress, which simplifies it a little.
Also introduce a new function to find games by player name, and make sure
we don't try to obtain player names from a qGoBoard before we have game data.