Change the way we render stones on the board
Get rid of stone QPixmapItems, instead render a single QPixmap with shadows,
stones, and (as we already did) marks.

The goal here is to be able, for exporting purposes, to render a board to a
pixmap of predefined size rather than having to rely on a screen grab.  Needs
a bit more work to also include lines and coordinates in prerendered pixmaps.