Create part of the View menu in the ui file, and move to a dock-based layout
Setting up desirable layouts with all the splitters became too unwieldy, and
dock views are made for just this problem.  However, they are not without
issues:
  - we have multiple different combinations of visible panes depending on
    game mode.  Qt's restoreState/restoreGeometry assume there is only one.
  - Qt has some bugs in this area, which can cause dock sizes to snap back
    to previous values sometimes.

Fortunately, the second problem has a workaround, and for the first, we now
keep track of a default window layout, and specific layouts for combinations
of visible panes.  We then add a bit of extra code to make sure things like
the observer list aren't shown in offline board windows.

To make things easier for users, add menu items to create some default
layouts.