Fix keyboard handling in clientwin
It was set up to let the command line have focus, even the escape key had
that effect.  The problem with this is that menu shortcuts don't work.
The reverse behaviour is more friendly and also consistent with the board
window, so change it.  Now there's no need for keyPressEvent any more.