Make MainWindow responsible for making moves and checking ko.
The game_state is the wrong place to check for ko - it doesn't know about the
ruleset, and a it does happen that a game contains illegal ko captures.
A new function MainWindow::make_move is used to perform a move and return
status information.  player_move is changed to call it.