Clear sendbuffer when going offline
Just guessing, but this might fix issue #15. I don't really have any
idea what happened in that issue, but I managed to confuse a local
NNGS to not send ready messages, leading to commands queueing up in
the send buffer and being sent to the server only after disconnect/reconnect.
Somehow that ended up with the server not registering the client setting,
and the logs in that issue suggest the same thing happened. I've not seen
this behaviour on IGS, but it may be possible to trigger it somehow.