Remove the game_state observer code
This seemed like a good idea at the time, but it turns out it's a complication
that does not buy us very much.  It was used for two purposes: to change a
board position in response to a move or an edit being made, and to keep track
of the current game position in a qGoBoard for an online game.  In both cases
we can just keep track of it manually, which is probably clearer anyway.