Simplify the code for players/games tables
There's one intended change in behaviour: the tables are not cleared when
sending games/who commands.  Instead, the existing entries are marked as not
up-to-date, and when the end of the list is reached, any who have not been
updated or newly created are removed.

There was all sorts of other complicated code which was removed, probably
without ill effects.  How to keep the lists sorted properly when not in quiet
mode remains an open problem.