Disable game tree header with Qt < 5.10
It seems Qt 5.9 is still fairly widespread (issue #22), so it's better to just
disable a non-critical feature rather than require a higher version.

I tried reverting 7c58b67d, which added the current header view and replaced
the one based on QHeaderView, but it now doesn't work for me even under Linux.
The headers are too wide, an issue I previously was seeing only on Windows.