Disentangle parse_move a bit: move out code for resume own game after adjourn
This function is a mess. To signal that a game is being resumed, parser.cpp
would call slot_gamemove rather than slot_matchcreate as one might expect,
and signal the special property of the call by setting the game's size to "@@".
There is code to look for and reuse an existing window with the adjourned game,
but it's outdated and never triggers because we remove adjourned games from
the list.

This is now improved with the addition of a new resume_own_game method.
Unfortunately this is untested. Last time I tried, IGS wouldn't even let me
adjourn games ("That command is currently disabled").