Reimplement table views using QTreeView/QAbstractItemModel
The existing code was a quick port of the old Qt3 stuff, which means it
was pretty awful and had weird performance issues every now and then.

To reduce CPU consumption when using non-quiet mode, changes to players and
games are now batched up and the view is updated only once a second.

This also includes improvements to filtering, the combo boxes now respond
immediately.