Remember quiet mode setting per-host
This is issue #36.

This is perhaps more complicated than strictly necessary. It seems
like at least on IGS this is saved server-side along with open and
looking, but we still want to keep a local setting so we can choose a
default.  For IGS, we want quiet on by default, since quiet off
doesn't behave all that well (the server ignores the current room, and
apparently black/white are somewhat random when a game is started).