Remove some unnecessary code from Talk dialogs
This ended up appending unwanted newlines.
Archeology is really unclear as to what the original purpose of this code was,
it morphed a bit over time when adapting to new versions of Qt.

In an earlier CVS revision of qGo, this was

  // set cursor to last line
  dlg->get_mle()->setCursorPosition(dlg->get_mle()->numLines(), 999);
  dlg->get_mle()->insertLine("");
  dlg->get_mle()->removeLine(dlg->get_mle()->numLines()-2);

which makes very little sense to me. But very probably there is no need for
any of this anymore if there ever was.