Change how threading works in the search
The previous setup was to search the symmetries of a pattern independently,
which was easy to do and worked reasonably well.
Now, we split the search in batches of games, examining each game only
once. This reduces the amount of time spent loading move lists, allows for
threading past 8 cores, and makes some other planned changes easier.