Don't crash when analyzing games with stupid komi values
Given an SGF file with komi values of, say 650, KataGo will complain.  This
causes q5go to become confused, thinking there was a fatal error.  In this
state KataGo can still send analysis lines, but q5go no longer expects them
and crashes.

This is a minimal fix. It could be improved by reporting the issue to the
user, through a popup or maybe by coloring the displayed komi in the sidebar
red.