Add option to show variations as score difference
Formerly, the UI offered to show variations as win rate, win rate
difference, and absolute score. Add the option to show variations as
score difference.

This is particularly helpful for analyzing amateur games where the
winrate as computed by katago assumes perfect play which is unrealistic.
The score difference gives more information on how well amateur players
are doing with each stone.

Additionally, when variations are shown as absolute score or score
difference, use a color scheme based on the score difference rather than
based on the winrate difference. It is confusing when colors do not
match the numbers.