always return an analysis for the next move
This way, even if the next move had not been considered by kata-analyze,
we can display a winrate/score estimation based on the analysis of the
next board position. This allows the user to see at a glance, through
render_analysis_marks (), how bad the move they actually played was.